+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     - Add documentation for filtering to `document_index.ipynb`.
 - Add `StudioClient` as a connector for submitting traces.
 - You can now specify a `chunk_overlap` when creating an index in the Document Index.
+- Add support for monitoring progress in the document index connector when embedding documents.
 
 ### Fixes
  - TaskSpan now properly sets its status to `Error` on crash.